# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Silvasonic Status Board",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/devices": {
      "get": {
        "tags": [
          "Management API"
        ],
        "summary": "List Devices",
        "description": "List all devices with optional filtering.",
        "operationId": "list_devices_api_v1_devices_get",
        "parameters": [
          {
            "name": "enrollment_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(pending|enrolled|ignored)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Enrollment Status"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceResponse"
                  },
                  "title": "Response List Devices Api V1 Devices Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/devices/{serial_number}": {
      "get": {
        "tags": [
          "Management API"
        ],
        "summary": "Get Device",
        "description": "Get single device details.",
        "operationId": "get_device_api_v1_devices__serial_number__get",
        "parameters": [
          {
            "name": "serial_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Serial Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Management API"
        ],
        "summary": "Update Device",
        "description": "Update device state (Enroll, Ignore, etc).",
        "operationId": "update_device_api_v1_devices__serial_number__patch",
        "parameters": [
          {
            "name": "serial_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Serial Number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/profiles": {
      "get": {
        "tags": [
          "Management API"
        ],
        "summary": "List Profiles",
        "description": "List all microphone profiles.",
        "operationId": "list_profiles_api_v1_profiles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProfileResponse"
                  },
                  "type": "array",
                  "title": "Response List Profiles Api V1 Profiles Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Management API"
        ],
        "summary": "Create Profile",
        "description": "Create a new custom profile.",
        "operationId": "create_profile_api_v1_profiles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/profiles/{slug}": {
      "delete": {
        "tags": [
          "Management API"
        ],
        "summary": "Delete Profile",
        "description": "Delete a profile (non-system only).",
        "operationId": "delete_profile_api_v1_profiles__slug__delete",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Profile Api V1 Profiles  Slug  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/containers/{container_id}/{action}": {
      "post": {
        "tags": [
          "Management API"
        ],
        "summary": "Control Container",
        "description": "Control a container (start/stop/restart).",
        "operationId": "control_container_api_v1_containers__container_id___action__post",
        "parameters": [
          {
            "name": "container_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Container Id"
            }
          },
          {
            "name": "action",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Action"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Control Container Api V1 Containers  Container Id   Action  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "DeviceResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Logical name of the device"
          },
          "enrollment_status": {
            "type": "string",
            "title": "Enrollment Status",
            "description": "pending, enrolled, or ignored"
          },
          "profile_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Slug",
            "description": "Linked microphone profile"
          },
          "serial_number": {
            "type": "string",
            "title": "Serial Number"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "last_seen": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Last Seen"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          }
        },
        "type": "object",
        "required": [
          "name",
          "enrollment_status",
          "serial_number",
          "model",
          "status",
          "last_seen",
          "enabled",
          "config"
        ],
        "title": "DeviceResponse",
        "description": "Full Device response."
      },
      "DeviceUpdate": {
        "properties": {
          "enrollment_status": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(pending|enrolled|ignored)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enrollment Status"
          },
          "profile_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Slug"
          },
          "logical_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logical Name"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          }
        },
        "type": "object",
        "title": "DeviceUpdate",
        "description": "Payload for updating a device state."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ProfileCreate": {
        "properties": {
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9-_]+$",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "match_pattern": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Pattern"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "name",
          "config"
        ],
        "title": "ProfileCreate",
        "description": "Payload for creating a new profile."
      },
      "ProfileResponse": {
        "properties": {
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9-_]+$",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "match_pattern": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Pattern"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_system": {
            "type": "boolean",
            "title": "Is System"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "name",
          "config",
          "is_system"
        ],
        "title": "ProfileResponse",
        "description": "Full Profile response."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}